control: Add build-dep on libnss-myhostname
The HSTS tests are failing because of not being able to resolve
subdomains of localhost. If the libnss-myhostname package is installed
while the tests are run, then the name resolution should work correctly.

See also https://gitlab.gnome.org/GNOME/libsoup/issues/159

Thanks: Dan Nicholson for figuring out the problem.