libsoup-gnome2.4-dev: Explicitly depend on gir1.2-soup-2.4
According to the GIR mini-policy, this is required because
gir1.2-soup-2.4 contains SoupGNOME-2.4.typelib, corresponding to
SoupGNOME-2.4.gir in libsoup-gnome2.4-dev. This dependency is not in
fact strictly necessary, because libsoup-gnome2.4-dev depends on
libsoup2.4-dev which in turn depends on gir1.2-soup-2.4, but Lintian
doesn't look at recursive dependencies.